data: Convert update_databases.py to Python-3.

+++ data/update_databases.py
@@ -1,5 +1,7 @@
+#!/usr/bin/env python3
+
 import pickle
-import urllib2
+import urllib.request, urllib.error, urllib.parse
 import time 
 import gzip
 import shutil
@@ -9,16 +11,22 @@ import os
 
 from bs4 import BeautifulSoup
 
-print("Archiving old database")
-old_version = open('VERSION').readline().strip()
-os.mkdir(old_version + '.archive')
-for f in [x for x in os.listdir('.') if x.endswith('.pickle')]:
-    shutil.move(f, os.path.join(old_version + '.archive', f))
+version_file = 'VERSION'
+if os.path.exists(version_file):
+    print("Archiving old database ..")
+    with open(version_file) as f:
+        old_version = f.readline().strip()
+    print(("Found old version {}".format(old_version)))
+    os.mkdir(old_version + '.archive')
+    for f in [x for x in os.listdir('.') if x.endswith('.pickle')]:
+        shutil.move(f, os.path.join(old_version + '.archive', f))
 
+print("Writing new version file ..")
 date=time.strftime("%d-%m-%Y")
-with open('VERSION', 'w') as o:
-    o.write(date+'\n')
-print("Done")
+version=date
+with open(version_file, 'w') as o:
+    o.write(version+'\n')
+print(("Done, wrote version {}".format(version)))
 
 
 output_dict             = {}
@@ -57,7 +65,7 @@ r2rclass_result         = 'reaction_to_rpair.%s.pickle' % date
 
 def build_name_dict(url):
     output_dictionary = {}
-    for entry in urllib2.urlopen(url).read().strip().split('\n'):
+    for entry in urllib.request.urlopen(url).read().decode('utf8').strip().split('\n'):
         sentry = entry.split('\t')
         key = sentry[0].split(':')[1]
         item = sentry[1].split(';')[0]
@@ -66,7 +74,7 @@ def build_name_dict(url):
 
 def build_dict(url):
     output_dictionary = {}
-    for entry in urllib2.urlopen(url).read().strip().split('\n'):
+    for entry in urllib.request.urlopen(url).read().decode('utf8').strip().split('\n'):
         key, item = [x.split(':')[1] for x in entry.split('\t')]
         if key in output_dictionary:
             output_dictionary[key].append(item)
@@ -83,10 +91,10 @@ def build_pfam_dict(url):
     clan2pfam           = {}
     with tempfile.NamedTemporaryFile(prefix='for_file', suffix='.gz') as tmp:
         cmd = 'wget -O %s %s' % (tmp.name, url)
-        subprocess.call(cmd, shell=True)
+        subprocess.check_call(cmd, shell=True)
         for line in gzip.open(tmp.name):
             pfam, clan, clan_name, pfam_name, description \
-                = line.strip().split('\t')
+                = line.decode('utf8').strip().split('\t')
             if clan.startswith('CL') and len(clan)==6:
                 pfam2clan[pfam] = clan
                 clan2name[clan] = clan_name
@@ -101,82 +109,81 @@ def build_pfam_dict(url):
 print("Downloading PFAM clan information")
 pfam2clan, clan2name, pfam2name, pfam2description, clan2pfam = build_pfam_dict(PFAM2CLAN)
 print("Done")
-print("Pickling results: %s" % pfam2clan_result)
+print(("Pickling results: %s" % pfam2clan_result))
 pickle.dump(pfam2clan, open(pfam2clan_result, "wb"))
-print("Pickling results: %s" % clan2name_result)
+print(("Pickling results: %s" % clan2name_result))
 pickle.dump(clan2name, open(clan2name_result, "wb"))
-print("Pickling results: %s" % pfam2name_result)
+print(("Pickling results: %s" % pfam2name_result))
 pickle.dump(pfam2name, open(pfam2name_result, "wb"))
-print("Pickling results: %s" % pfam2description_result)
+print(("Pickling results: %s" % pfam2description_result))
 pickle.dump(pfam2description, open(pfam2description_result, "wb"))
-print("Pickling results: %s" % clan2pfam_result)
+print(("Pickling results: %s" % clan2pfam_result))
 pickle.dump(clan2pfam, open(clan2pfam_result, "wb"))
-exit()
 
 print("Downloading reaction to orthology information from KEGG")
 r2k = build_dict(R2K)
 print("Done")
-print("Pickling results: %s" % r2k_result)
+print(("Pickling results: %s" % r2k_result))
 pickle.dump(r2k, open(r2k_result, "wb"))
 
 print("Downloading reaction to pathway information from KEGG")
 r2p = build_dict(R2P)
 print("Done")
-print("Pickling results: %s" % r2p_result)
+print(("Pickling results: %s" % r2p_result))
 pickle.dump(r2p, open(r2p_result, "wb"))
 
 print("Downloading pathway to reaction information from KEGG")
 p2r = build_dict(P2R)
 print("Done")
-print("Pickling results: %s" %  p2r_result)
+print(("Pickling results: %s" %  p2r_result))
 pickle.dump(p2r, open(p2r_result, "wb"))
 
 print("Downloading reaction to module information from KEGG")
 r2m = build_dict(R2M)
 print("Done")
-print("Pickling results: %s" % r2m_result)
+print(("Pickling results: %s" % r2m_result))
 pickle.dump(r2m, open(r2m_result, "wb"))
 
 print("Downloading reaction to module information from KEGG")
 m2r = build_dict(M2R)
 print("Done")
-print("Pickling results: %s" % m2r_result)
+print(("Pickling results: %s" % m2r_result))
 pickle.dump(m2r, open(m2r_result, "wb"))
 
 print("Downloading reaction to compound information from KEGG")
 r2c = build_dict(R2C)
 print("Done")
-print("Pickling results: %s" % r2c_result)
+print(("Pickling results: %s" % r2c_result))
 pickle.dump(r2c, open(r2c_result, "wb"))
 
 print("Downloading compound to reaction information from KEGG")
 c2r = build_dict(C2R)
 print("Done")
-print("Pickling results: %s" % c2r_result)
+print(("Pickling results: %s" % c2r_result))
 pickle.dump(c2r, open(c2r_result, "wb"))
 
 print("Downloading compound descriptions from KEGG")
 c   = build_name_dict(C)
 print("Done")
-print("Pickling results: %s" % c_result)
+print(("Pickling results: %s" % c_result))
 pickle.dump(c, open(c_result, "wb"))
 
 print("Downloading pathway descriptions from KEGG")
 p   = build_name_dict(P)
 print("Done")
-print("Pickling results: %s" % p_result)
+print(("Pickling results: %s" % p_result))
 pickle.dump(p, open(p_result, "wb"))
 
 print("Downloading reaction descriptions from KEGG")
 r   = build_name_dict(R)
 print("Done")
-print("Pickling results: %s" % r_result)
+print(("Pickling results: %s" % r_result))
 pickle.dump(r, open(r_result, "wb"))
 
 print("Downloading module descriptions from KEGG")
 m   = build_name_dict(M)
 print("Done")
-print("Pickling results: %s" % m_result)
+print(("Pickling results: %s" % m_result))
 pickle.dump(m, open(m_result, "wb"))
 
 
@@ -188,7 +195,7 @@ B='B'
 C='C'
 D='D'
 
-for line in urllib2.urlopen(br08001).read().strip().split('\n'):
+for line in urllib.request.urlopen(br08001).read().decode('utf8').strip().split('\n'):
     if line.startswith(A):
         A_text = list(BeautifulSoup(line, "html.parser").find('b').stripped_strings)[0]
     elif line.startswith(B):
@@ -215,14 +222,14 @@ for line in urllib2.urlopen(br08001).read().strip().split('\n'):
             output_dict[CID][C].append(C_text)
             output_dict[CID][D].append(D_text)
 
-print("Pickling results: %s" % output_pickle)
+print(("Pickling results: %s" % output_pickle))
 pickle.dump(output_dict, open(br08001_result, "wb"))
 
 print("Done")
 
 
 
-current_module_list = m.keys()
+current_module_list = list(m.keys())
 output_pickle = 'module_to_definition.%s.pickle' % date
 print("Downloading module definitions from KEGG")
 m2def={}
@@ -231,25 +238,25 @@ tot = float(len(current_module_list))
 for idx, module_key in enumerate(current_module_list):
     url = base + module_key
     m2def[module_key]=[]
-    for line in urllib2.urlopen(url).read().strip().split('\n'):
+    for line in urllib.request.urlopen(url).read().decode('utf8').strip().split('\n'):
         if line.startswith('DEFINITION'):
             m2def[module_key] = ' '.join(line.split()[1:])
-    print "%s percent done" % str(round(float(idx+1)/tot, 2)*100)
+    print(("%s percent done" % str(round(float(idx+1)/tot, 2)*100)))
     
 print("Done")
-print("Pickling results: %s" % output_pickle)
+print(("Pickling results: %s" % output_pickle))
 pickle.dump(m2def, open(output_pickle, "wb"))
 
 
 print("Downloading rpair information from KEGG")
 r2rclass={}
 base='http://rest.kegg.jp/get/'
-tot = float(len(r.keys()))
+tot = float(len(list(r.keys())))
 for idx, reaction_key in enumerate(r.keys()):
     url = base + reaction_key
     r2rclass[reaction_key]=[]
     try:
-        for line in urllib2.urlopen(url).read().strip().split('\n'):
+        for line in urllib.request.urlopen(url).read().strip().split('\n'):
             if line.startswith('RCLASS'):
                 r2rclass[reaction_key]+=line.split()[2:]
                 
@@ -258,9 +265,10 @@ for idx, reaction_key in enumerate(r.keys()):
                 if sline[0].startswith('RC'):
                     r2rclass[reaction_key]+=sline[1:]
     except:
-        print reaction_key
-    print "%s percent done" % str(round(float(idx+1)/tot, 2)*100)
+        print(reaction_key)
+    print(("{} percent done".format(str(round(float(idx+1)/tot, 2)*100))))
     
 print("Done")
-print("Pickling results: %s" % m_result)
-pickle.dump(r2rclass, open(r2rclass_result, "wb"))
+print(("Pickling results: {}".format(m_result)))
+with open(r2rclass_result, "wb") as f_out:
+    pickle.dump(r2rclass, f_out)
